📦 feat (work details): finish work saving and listing func

+++ app/src/main/java/com/kabirnayeem99/paymentpaid/utils/CustomUtils.kt
@@ -19,7 +19,7 @@ object CustomUtils {
      * @param month which is an [Int]
      * @return month which is a [String]
      */
-    fun padMonth(month: Int): String {
+    fun padMonth(month: Long): String {
         val output = StringBuilder(month.toString())
         if (output.length < 2) {
             output.insert(0, "0")
